test(automation): Add GitHub action for testing
Automatically runs Django tests on every push to the staging branch, &
depending upon the result, it acts differently. If the tests fail, it
creates a pull request with the title "[Automated] Fix failing tests"
to review and fix the problems. If the tests pass, it merges the
staging branch into the main branch and runs build commands
for npm. Overall, this workflow helps to automate the testing process,
making it more efficient and reliable.