feat: add subscription form to home page
- Added a subscription form to the home view. It handles the
post-request. If the form is valid and there are no existing
subscriptions with the provided email, the subscription is saved and
the user is redirected back to the home page. If the form is not valid,
the errors are added to the context and sent back to the home view.