feat: add 'Open Sans' font & enhanced form styles
Added multiple weights of the 'Open Sans' font for better typography
across the site. Also, did various enhancements in the form styling,
which includes better placement of form elements, styling for alert
messages, addition of privacy policy details and newly improved design
for select elements for better UX. This should improve the visual
identity and user experience of the site.