refactor: CustomerProfile to DeliveryDetails model
Renamed the "CustomerProfile" model to "DeliveryDetails" to better
reflect the data it stores. The change includes adding new fields for
default full name and email, which were not present earlier. The verbose
name for the model is also updated to 'Delivery Details' to ensure
consistency. All instances of CustomerProfile have been updated to
DeliveryDetails in the save method.