refactor: forms in accounts app
Modified forms.py in the accounts app to better cater to delivery
details. The CustomLoginForm was changed to UserUpdateForm, with
placeholder & help_text now added for each field.  Also, the
CustomerProfileForm was transformed to CustomerDeliveryForm to better
represent the usage of the form. These changes aim to make UX more
intuitive and the code more maintainable.