feat: add wishlist functionality
Update includes adjustments to the layout and styles, a Wishlist feature
is added using JavaScript to display a wishlist dialog so logged-in
users can add wishlist products from the cart.

Updated cart styles for mobile views to prevent overflow issues from
tables views. A ‘continue shopping’ button is added when the cart is
empty to redirect users back to the product page.

issue: #16