feat: add wishlist delete functionality
Added the ability for users to delete their wishlists.
New imports were added to support this functionality, and a
new DeleteView was created to handle the wishlist deletion process. The
Django's get_object_or_404 was used to ensure that the wishlist to be
deleted exists, and it belongs to the current user. A success message
is displayed after successful deletion.

issue: #16