feat: add email confirmation functionality in signup
1. Added a ConfirmSignupView in the subscriptions app that confirms a
user's email once they click on the confirmation link sent to them. The
view checks for a valid confirmation_token in the Subscription model and
updates the respective record by setting is_verified to True, setting
the confirmed_at field to the current time, and subsequently removing
the confirmation_token. This will prevent the re-use of the
confirmation link. This was necessary to ensure user accounts that sign
up for the newsletter are valid and controlled by the user.

issue#15