feat: add stock control to cart update process
Whenever a user attempts to add or modify the quantity of a
product in the cart, the system will validate the
availability of the requested quantity in stock.

In cases where the stock is inadequate, an error message will be
displayed, and no modifications will be made to the cart. This
implementation effectively resolves the problem of users adding
unavailable quantities to their carts.

issue:#3