feat: add referring_url to form input
Added hidden input field in the wishlist form. This field captures and
submits the referring_url data, i.e., the current path.

The purpose of this change is to allow the backend to keep track of
which page led the user to add a product to their wishlist.
This feature can be essential for navigating the user back to the
original page after a product was successfully added to the users
wishlist.