feat: add overlay and close search bar on scroll or on keydown Escape
A div overlay was added to improve UX when using the search bar. The
overlay is visible when the search bar is active and will darken the
remaining part of the website, keeping focus on the search. Also
included functionality to close the search bar on various interactions -
scrolling and pressing the Escape key.