fix(HTML validation error): update email link
The hyperlink in the customer support template was updated. This was
done to ensure that the message body in the 'mailto' link is properly
URL encoded, avoiding possible issues with some email clients. The link
now includes '%2C%20' instead of a comma and space, which aligns with
the standard URL encoding practices.