feat: improve accessibility by adding ARIA labels and altering alt text
Added several ARIA labels to interactive elements in the
navigation menu, improving the website's accessibility for those using
screen readers. Now the purpose of each button or link is clearly
communicated.

Additionally, the alt text of certain images is updated to
better describe the image's content or function. The img tags flagged
as `aria-hidden="true"` are decorative and don't require alt text for
 screen readers.